refactor(chaos): 重构混沌引擎以支持独立上下文和决策结果
- 将混沌引擎重构为使用独立的 ChaosContext 和 DecisionResult 结构
- 添加 BuildUserPromptFromKernel 方法以适配内核上下文调用
- 在 auto_trader_chaos.go 中实现完整的混沌交易循环
- 修复 diagnose_orders.go 中的类型转换问题
- 更新类型定义以包含新的数据字段和决策结果结构

+++ scripts/diagnose_orders.go
@@ -121,7 +121,7 @@ func main() {
 
 		filledAtStr := "N/A"
 		if order.FilledAt > 0 {
-			filledAtStr = time.UnixMilli(order.FilledAt).Format("01-02 15:04")
+			filledAtStr = time.UnixMilli(int64(order.FilledAt)).Format("01-02 15:04")
 		}
 
 		fmt.Printf("%-15s %-10s %-10s %-15.2f %-10s %s\n",
